+++ feat_fast/getFocus/getFocus.hpp
@@ -86,7 +86,7 @@ public:
 };
 
 
-double getWarpFocal(const vector<float>& HomogMat, 
+double getWarpFocal(const vector<double>& HomogMat, 
 	size_t img1Size, size_t img2Size);
 
 void getWarpOffset(const ImgRaw &imgA, const ImgRaw &imgB,
